pyproject.toml
[project]
name = "tinyvla"
version = "0.1.0"
description = "Minimal Vision-Language-Action model in one file"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.0",
    "torchvision",
    "transformers>=4.45",
    "accelerate>=0.25",
    "peft>=0.7",
    "pillow",
    "tqdm",
    "numpy",
    "qwen-vl-utils",
    "datasets",
    "wandb",
]

[project.optional-dependencies]
eval = [
    "libero @ git+https://github.com/Lifelong-Robot-Learning/LIBERO.git",
    "robosuite>=1.4",
    "imageio",
    "imageio-ffmpeg",
]

full = [
    "tinyvla[eval]",
    "wandb",
    "bitsandbytes",
    "flash-attn",
]

dev = [
    "pytest",
    "ruff",
]

[project.scripts]
tinyvla = "tinyvla:main"

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
